BackEnd/Domains/Users/ListUser/requirements.txt
flasgger==0.9.7.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
flask-swagger-ui==4.11.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
gunicorn==20.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0

